requirements.txt
autopep8==1.5.7
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
decorator==5.0.9
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
pycodestyle==2.7.0
Pygments==2.9.0
pyxll==5.1.1
pyzmq==22.1.0
requests==2.25.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
websockets==9.1
